requirements.txt
django
djangorestframework
markdown
django-filter
djangorestframework-simplejwt
pytest
pytest-django
mixer